refactor: :hammer: refactor the project structure
Refactored the project structure to group related classes
