+++ src/Contracts/Authenticatable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace AndrewDyer\Gate;
+namespace AndrewDyer\Gate\Contracts;
 
 /**
  * Represents an entity that can be authenticated and used as an actor within the gate.
